+++ core-bundle/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Contao core bundle change log
 
+## DEV
+
+ * Correctly generate the login target URLs in the back end (see #1432).
+
 ## 4.5.6 (2018-03-06)
 
  * Use the DCA information to determine the index length (see contao/installation-bundle#88).